+++ CHANGELOG.md
@@ -1,2 +1,7 @@
 ### v1.0.0
 * Initial release
+
+### v1.0.1 (unreleased)
+* Overloaded Capacitor
+    * First stack damage bonus: 20% -> 15%
+* Added Relic Guard